Unified check_solution functions + renaming
