+++ pySDC/projects/PinTSimE/estimation_check_2capacitors.py
@@ -4,13 +4,12 @@ from pathlib import Path
 
 from pySDC.helpers.stats_helper import get_sorted
 from pySDC.core.Collocation import CollBase as Collocation
-from pySDC.implementations.problem_classes.Battery import battery_n_condensators
+from pySDC.implementations.problem_classes.Battery import battery_n_capacitors
 from pySDC.implementations.sweeper_classes.imex_1st_order import imex_1st_order
 from pySDC.implementations.controller_classes.controller_nonMPI import controller_nonMPI
 from pySDC.projects.PinTSimE.battery_model import get_recomputed
-from pySDC.projects.PinTSimE.battery_2condensators_model import get_data_dict
 from pySDC.projects.PinTSimE.piline_model import setup_mpl
-from pySDC.projects.PinTSimE.battery_2condensators_model import log_data, proof_assertions_description
+from pySDC.projects.PinTSimE.battery_2capacitors_model import check_solution, log_data, proof_assertions_description
 import pySDC.helpers.plot_helper as plt_helper
 
 from pySDC.projects.PinTSimE.switch_estimator import SwitchEstimator
@@ -73,7 +72,7 @@ def run(dt, use_switch_estimator=True):
 
     # fill description dictionary for easy step instantiation
     description = dict()
-    description['problem_class'] = battery_n_condensators  # pass problem class
+    description['problem_class'] = battery_n_capacitors  # pass problem class
     description['problem_params'] = problem_params  # pass problem parameters
     description['sweeper_class'] = imex_1st_order  # pass sweeper
     description['sweeper_params'] = sweeper_params  # pass sweeper parameters
@@ -274,67 +273,6 @@ def check(cwd='./'):
 
     fig2.savefig('data/diffs_estimation_vC2.png', dpi=300, bbox_inches='tight')
     plt_helper.plt.close(fig2)
-
-
-def check_solution(stats, dt, use_switch_estimator):
-    """
-    Function that checks the solution based on a hardcoded reference solution. Based on check_solution function from @brownbaerchen.
-
-    Args:
-        stats (dict): Raw statistics from a controller run
-        dt (float): initial time step
-        use_switch_estimator (bool): flag if the switch estimator wants to be used or not
-    """
-
-    data = get_data_dict(stats, use_switch_estimator)
-
-    if use_switch_estimator:
-        msg = f'Error when using the switch estimator for battery_2condensators for dt={dt:.1e}:'
-        if dt == 4e-1:
-            expected = {
-                'cL': 1.1842780233981391,
-                'vC1': 1.0094891393319418,
-                'vC2': 1.00103823232433,
-                'switch1': 1.6075867934844466,
-                'switch2': 3.209437912436633,
-                'restarts': 2.0,
-                'sum_niters': 2000,
-            }
-        elif dt == 4e-2:
-            expected = {
-                'cL': 1.180493652021971,
-                'vC1': 1.0094825917376264,
-                'vC2': 1.0007713468084405,
-                'switch1': 1.6094074085553605,
-                'switch2': 3.209437912440314,
-                'restarts': 2.0,
-                'sum_niters': 2364,
-            }
-        elif dt == 4e-3:
-            expected = {
-                'cL': 1.1537529501025199,
-                'vC1': 1.001438946726028,
-                'vC2': 1.0004331625246141,
-                'switch1': 1.6093728710270467,
-                'switch2': 3.217437912434171,
-                'restarts': 2.0,
-                'sum_niters': 8920,
-            }
-
-    got = {
-        'cL': data['cL'][-1],
-        'vC1': data['vC1'][-1],
-        'vC2': data['vC2'][-1],
-        'switch1': data['switch1'],
-        'switch2': data['switch2'],
-        'restarts': data['restarts'],
-        'sum_niters': data['sum_niters'],
-    }
-
-    for key in expected.keys():
-        assert np.isclose(
-            expected[key], got[key], rtol=1e-4
-        ), f'{msg} Expected {key}={expected[key]:.4e}, got {key}={got[key]:.4e}'
 
 
 if __name__ == "__main__":
