+++ pySDC/tests/test_projects/test_pintsime/test_battery_2capacitors_model.py
@@ -1,0 +1,8 @@
+import pytest
+
+
+@pytest.mark.base
+def test_main():
+    from pySDC.projects.PinTSimE.battery_2capacitors_model import main
+
+    main()
