--- pySDC/tests/test_projects/test_pintsime/test_battery_2condensators_model.py
@@ -1,8 +1,0 @@
-import pytest
-
-
-@pytest.mark.base
-def test_main():
-    from pySDC.projects.PinTSimE.battery_2condensators_model import main
-
-    main()
