+++ pySDC/tests/test_projects/test_pintsime/test_estimation_check_2capacitors.py
@@ -1,0 +1,8 @@
+import pytest
+
+
+@pytest.mark.base
+def test_main():
+    from pySDC.projects.PinTSimE.estimation_check_2capacitors import check
+
+    check()
