--- pySDC/tests/test_projects/test_pintsime/test_estimation_check_extended.py
@@ -1,8 +1,0 @@
-import pytest
-
-
-@pytest.mark.base
-def test_main():
-    from pySDC.projects.PinTSimE.estimation_check_extended import check
-
-    check()
